HTTP cleanup old structures
