+++ Sources/HAP/Endpoints/identify().swift
@@ -4,6 +4,6 @@ import HTTP
 func identify(device: Device) -> Responder {
     return { context, request in
         device.delegate?.didRequestIdentification()
-        return HTTPResponse(status: .noContent)
+        return .noContent
     }
 }
