--- Sources/HAP/HTTP/Application.swift
@@ -1,4 +1,0 @@
-//import Foundation
-//import KituraNet
-//
-//typealias Application = (Server.Connection, ServerRequest) -> Response
